Made a few changes in a session with Lydia to improve usability and make the sorting tool most useful to her work. The main changes were:
- changed the color of spike markers to be more visible in spike detection plot

- fixed the initial spike detection plots to all have the same y axis range
so they can be compared better by eye

- fixed a problem where, if the number of clusters was from the beginning
less or equal change_clusters, spikes that were not assign to clusters initially
(because they belonged to a too small cluster) remained unassigned (unit -1).
Now, these are assigned to the closest matching cluster even when the number of
clusters is smaller than change_clusters

- introduced a manual merge approval mechanism: if in the "spike sort" section
approve_merge is set to True, the user is asked to agree to every cluster merge
explicitly. If approval is refused, the merge is permanently banned in an
illegal_merges list and is never suggested again.