Fixed a problem where we were using teh wrong information for the spike rate. Note to self: Also should check how spikerates are estimated as it seems they have a little bit of a delay of going up after a couple of quick spikes.
Also now found a reference to teh memory leak problem which occurs in gui
backends but not non-gui such as agg.
Allegedly, also does not occur in gui backends if figures are shown and then
closed. This does not seem to be tha case for me
(see https://github.com/matplotlib/matplotlib/issues/20300 for reference)